add toggles for individual Phantoon patterns (#62)
Add hooks to Phantoon's AI that override the default RNG-based
pattern chooser in order to limit patterns to user-selected subset.

If the user enabled all the patterns (or none of them), Phantoon's
regular RNG code is run instead -- so this feature should not impact
Phantoon's behavior or pattern probabilities when not in use.