Show ram_last_realtime_door in infohud
Modified the infohud to use ram_last_realtime_door when "Transition Lag"
is set to "Full".  Previously, with "Transition Lag" set to "Full" the
practice hack would set ram_last_door_lag_frames to the value to be
displayed.

With this change, ram_last_door_lag_frames only counts the lag frames in
the transition, even with "Transition Lag" set to "Full".  This is
important for programs that scrape the practice hack's memory, so they
get the same count for door lag regardless of what the user selected to
be displayed in the infohud.