Avoid accidental pauses from controller shortcuts
As recently discussed in Discord, controller shortcuts involving the
start button (most commonly the Start+Select menu shortcut) will cause
the game to pause unless start is the last input pressed. In order to
avoid this, check if the game is in the "fading out to pause" mode
after handling most shortcuts. If so, cancel the pause by returning to
the "normal gameplay" mode and resetting the screen brightness.

This behavior applies to all controller shortcuts except loading presets
(because presets reset the game mode anyway) and saving/loading
savestates (because the user might actually want to savestate in the
middle of a pause, for instance to practice a gravity jump).