[GH Actions] Upload build artifacts for PRs
When we build a PR, automatically upload the resulting .ips files to
GitHub as build artifacts. This is intended to make it easier to test
PRs without needing to manually check out and build the branch.

To avoid confusion, pull request builds get a "prerelease" note with the
Git commit hash in the main menu so that testing and release builds can
be easily distinguished. This is implemented using a `PRERELEASE` macro
that is not defined by default but is set to the Git commit short hash
in GitHub Actions. In order for this to work, I had to make a small
change to the build scriptts which passes arguments through to ASAR.

This change does not affect non-pull-request builds in any way.