Use fast decompression if fast doors are enabled
This fixes the Rinkas in Mother Brain's room spawning incorrectly. The
Rinkas try to spawn in order from left-to-right, but since (with slow
doors) the rinkas on the first zebetite are not visible when
decompression finishes and enemies are loaded, they fail to spawn and
the rinkas next to the door spawn instead. When the zeb is scrolled
onscreen, only the top rinka is spawned initially because only 3 rinkas
can be onscreen at once.

Under the previous iteration of fast doors, decompression finished later
in the door transition, causing the rinkas on the first zeb to spawn
early & messing up the timing for the downback zeb skip. Using fast
decompression causes the sequencing to behave like in vanilla.