Don't kill invisible enemies
The flame jets in LN Main Hall are two enemies. One part is intangible, meaning it will not be killed. The other part is invisible, but handles collision. We should keep the collision since the graphics are always there. I think this would have less side effects than deleting the intangible enemies.