Add timer mode that adds skipped cutscenes and doors
Removes the "Adjust Fanfare Timer" option. In its place, the
"Realtime/Ingame" timer selector gets a third option, called "Speedrun"
(could be renamed to "Vanilla" or "Adjusted" or something). This option
is real-time except as it would be during a run of the vanilla game, so
skipped item fanfares & boss intros, as well as fast doors and
elevators, have the skipped time added back in to the room and segment
timers.

Upon upgrade, users who had the "adjust fanfare" option turned on and
their timer set to realtime will automatically have their timer changed
to speedrun mode.