Add boot-to-menu functionality
Add an option to boot directly to the InfoHUD menu instead of the
Nintendo splash. The sound engine is also loaded asynchronously in the
menu, elimiating the delay for it to initialize. This reduces the time
from power-on to practice menu to about half a second.

The option is under the Cutscenes menu.